Add pull to refresh indicator for comments
