+++ feed/src/main/java/com/neaniesoft/warami/featurefeed/CommentsScreen.kt
@@ -1,59 +1,16 @@
 package com.neaniesoft.warami.featurefeed
 
 import android.util.Log
-import androidx.compose.foundation.clickable
-import androidx.compose.foundation.interaction.MutableInteractionSource
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.IntrinsicSize
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxHeight
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.layout.width
-import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.material.Text
-import androidx.compose.material.ripple.rememberRipple
-import androidx.compose.material3.Divider
-import androidx.compose.material3.Icon
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Surface
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
-import androidx.compose.runtime.remember
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.vector.rememberVectorPainter
-import androidx.compose.ui.layout.ContentScale
-import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.res.pluralStringResource
-import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
-import coil.compose.AsyncImage
-import coil.request.ImageRequest
-import com.neaniesoft.warami.common.extensions.formatPeriod
-import com.neaniesoft.warami.common.models.ChildCount
-import com.neaniesoft.warami.common.models.Comment
 import com.neaniesoft.warami.common.models.CommentId
 import com.neaniesoft.warami.common.models.PostId
-import com.neaniesoft.warami.common.models.Score
-import com.neaniesoft.warami.common.models.UriString
-import com.neaniesoft.warami.featurefeed.components.icons.CommentIcons
-import com.neaniesoft.warami.featurefeed.components.shapes.SpeechBubbleShape
+import com.neaniesoft.warami.featurefeed.components.comment.CommentsScreenContent
 import com.ramcosta.composedestinations.annotation.Destination
 import com.ramcosta.composedestinations.navigation.DestinationsNavigator
-import java.time.Instant
-import java.time.ZoneId
-import java.time.ZonedDateTime
 
 @Destination
 @Composable
@@ -65,6 +22,7 @@ fun CommentsScreen(
 ) {
     Log.d("CommentsScreen", "PostId: $postId")
     val commentsWithDepth by viewModel.comments.collectAsState()
+    val isRefreshing by viewModel.isRefreshing.collectAsState()
 
     LaunchedEffect(key1 = Unit) {
         Log.d("CommentsScreen", "refreshing $postId, parent: $parentCommentId")
@@ -79,226 +37,16 @@ fun CommentsScreen(
         }
     }
 
+    val pageLoadingContent by viewModel.pageLoadingContent.collectAsState()
+
     CommentsScreenContent(
+        postId = postId,
+        parentCommentId = parentCommentId,
         commentsWithDepth = commentsWithDepth,
+        pageLoadingContent = pageLoadingContent,
+        isRefeshing = isRefreshing,
         onLoadMoreCommentsClicked = viewModel::onLoadMoreCommentsClicked,
+        onNextPageClicked = { postId, commentId -> viewModel.onLoadNextPage(postId, commentId) },
+        onRefreshSelected = { postId, parentCommentId -> viewModel.refresh(postId, parentCommentId)}
     )
-}
-
-@Composable
-fun CommentsScreenContent(
-    commentsWithDepth: List<Pair<Comment, Int>>,
-    onLoadMoreCommentsClicked: (PostId, CommentId) -> Unit,
-) {
-    val maxDepth = remember(commentsWithDepth) { commentsWithDepth.maxByOrNull { it.second }?.second ?: 0 }
-    Surface(modifier = Modifier.fillMaxSize()) {
-        LazyColumn {
-            items(commentsWithDepth.size) { index ->
-                val (comment, depth) = commentsWithDepth[index]
-                CommentRow(
-                    postId = comment.postId,
-                    commentId = comment.commentId,
-                    creatorName = comment.creator.displayName ?: comment.creator.name,
-                    creatorAvatarUri = comment.creator.avatarUrl,
-                    score = comment.counts.score,
-                    time = comment.publishedAt.formatPeriod(
-                        LocalContext.current.resources, Instant.now().atZone(ZoneId.of("UTC")),
-                    ),
-                    body = comment.content,
-                    depth = depth,
-                    maxDepth = maxDepth,
-                    childCount = comment.counts.childCount,
-                    onLoadMoreCommentsClicked = onLoadMoreCommentsClicked,
-                )
-                val nextDepth = commentsWithDepth.getOrNull(index + 1)?.second
-                if (nextDepth != depth) {
-                    Divider()
-                }
-            }
-        }
-    }
-}
-
-data class CommentForDisplay(
-    val creatorName: String,
-    val creatorAvatarUri: UriString?,
-    val score: Score,
-    val time: ZonedDateTime,
-    val body: String,
-    val depth: Int,
-)
-
-@Composable
-fun CommentRow(
-    postId: PostId,
-    commentId: CommentId,
-    creatorName: String,
-    creatorAvatarUri: UriString?,
-    score: Score,
-    time: String,
-    body: String,
-    depth: Int,
-    maxDepth: Int,
-    childCount: ChildCount,
-    onLoadMoreCommentsClicked: (PostId, CommentId) -> Unit,
-) {
-    Row(
-        modifier = Modifier
-            .fillMaxWidth()
-            .height(IntrinsicSize.Min),
-    ) {
-        if (depth >= 1) {
-            repeat((1..depth).count()) {
-                val width = if (it == 0) {
-                    11.dp
-                } else {
-                    12.dp
-                }
-                Spacer(modifier = Modifier.width(width))
-                Divider(
-                    modifier = Modifier
-                        .width(1.dp)
-                        .fillMaxHeight(),
-                )
-            }
-        }
-
-
-        Surface(tonalElevation = 4.dp) {
-
-            Column(modifier = Modifier.fillMaxWidth()) {
-                Column(
-                    modifier = Modifier
-                        .padding(16.dp)
-                        .fillMaxWidth(),
-                ) {
-                    CommentHeader(creatorName = creatorName, creatorAvatarUri = creatorAvatarUri, score = score, time = time)
-                    Surface(
-                        modifier = Modifier.padding(top = 8.dp),
-                        tonalElevation = 16.dp,
-                        shadowElevation = 2.dp,
-                        shape = SpeechBubbleShape(triangleSize = 6.dp),
-                    ) {
-                        Text(
-                            text = body,
-                            style = MaterialTheme.typography.bodyMedium,
-                            modifier = Modifier.padding(start = 12.dp, end = 12.dp, top = 12.dp, bottom = 6.dp),
-                            color = MaterialTheme.colorScheme.onSurface,
-                        )
-
-                    }
-                }
-
-                if (depth == maxDepth && childCount.value > 0) {
-                    val interactionSource = remember {
-                        MutableInteractionSource()
-                    }
-                    Surface(
-                        modifier = Modifier
-                            .fillMaxWidth()
-                            .clickable(interactionSource = interactionSource, indication = rememberRipple(bounded = true)) {
-                                onLoadMoreCommentsClicked(postId, commentId)
-                            },
-                        tonalElevation = 16.dp,
-                    ) {
-                        Text(
-                            modifier = Modifier.padding(16.dp),
-                            text = pluralStringResource(id = R.plurals.load_more_comments, count = childCount.value, childCount.value),
-                            style = MaterialTheme.typography.labelMedium,
-                            color = MaterialTheme.colorScheme.onSurface,
-                        )
-                    }
-                }
-            }
-
-        }
-
-    }
-}
-
-@Composable
-@Preview
-fun CommentRowPreview() {
-    MaterialTheme {
-        Surface(modifier = Modifier.fillMaxWidth()) {
-            CommentRow(
-                postId = PostId(1),
-                commentId = CommentId(1),
-                creatorName = "Some person",
-                creatorAvatarUri = null,
-                score = Score(50),
-                time = "12 hours",
-                body = "Lorem ipsum dolor etc ueafh siueh fiushr giush ifuhage ifygsirgh iksh rfousoijf pdotijgpidjhrgkfhsge fuygsefy gsiurhg osr jglih eroughwiregf iwh grefkiuwh rogh woruhg ikwuehfiwug rigfh",
-                depth = 8,
-                maxDepth = 8,
-                childCount = ChildCount(3),
-                onLoadMoreCommentsClicked = { _, _ -> },
-            )
-        }
-    }
-}
-
-@Composable
-fun CommentHeader(
-    creatorName: String,
-    creatorAvatarUri: UriString?,
-    score: Score,
-    time: String,
-) {
-    Row(
-        modifier = Modifier.fillMaxWidth(),
-        horizontalArrangement = Arrangement.SpaceBetween,
-        verticalAlignment = Alignment.CenterVertically,
-    ) {
-        if (creatorAvatarUri == null) {
-            Icon(
-                modifier = Modifier.size(24.dp),
-                imageVector = CommentIcons.rememberPerson_2(),
-                contentDescription = stringResource(id = R.string.content_description_avatar_icon),
-            )
-        } else {
-            AsyncImage(
-                modifier = Modifier.size(24.dp),
-                model = ImageRequest.Builder(LocalContext.current)
-                    .crossfade(true)
-                    .data(creatorAvatarUri.value)
-                    .build(),
-                contentDescription = stringResource(id = R.string.content_description_avatar_icon),
-                contentScale = ContentScale.Crop,
-                placeholder = rememberVectorPainter(image = CommentIcons.rememberPerson_2()),
-            )
-        }
-        Text(
-            modifier = Modifier
-                .padding(start = 8.dp),
-            text = creatorName, style = MaterialTheme.typography.labelSmall, color = MaterialTheme.colorScheme.onSurface,
-        )
-        Spacer(modifier = Modifier.weight(1.0f))
-        Text(
-            text = score.value.toString(),
-            style = MaterialTheme.typography.labelSmall,
-            fontWeight = FontWeight.Bold,
-            color = MaterialTheme.colorScheme.onSurface,
-        )
-        Text(
-            modifier = Modifier.padding(start = 8.dp, end = 8.dp),
-            text = "•",
-            style = MaterialTheme.typography.labelSmall,
-            color = MaterialTheme.colorScheme.onSurface,
-        )
-        Text(text = time, style = MaterialTheme.typography.labelSmall, color = MaterialTheme.colorScheme.onSurface)
-
-    }
-}
-
-@Preview
-@Composable
-fun CommentHeaderPreview() {
-    MaterialTheme {
-        Surface(modifier = Modifier.fillMaxWidth()) {
-
-            CommentHeader(creatorName = "Some person", creatorAvatarUri = null, score = Score(50), time = "12 hours")
-
-        }
-    }
 }
